[Refactor] STAMPIT-130 필터링 섹션 레이아웃 extension 메서드로 추출
