+++ StampIt-Project/StampIt-Project/Resource/Extensions/NSCollectionLayoutSection+Extension.swift
@@ -1,0 +1,28 @@
+//
+//  NSCollectionLayoutSection+Extension.swift
+//  StampIt-Project
+//
+//  Created by daeun on 6/25/25.
+//
+
+import UIKit
+
+extension NSCollectionLayoutSection {
+    /// 칩 형태의 필터링 섹션 레이아웃
+    static func createFilterSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                              heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(0.2),
+                                               heightDimension: .absolute(32))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+
+        let section = NSCollectionLayoutSection(group: group)
+        section.interGroupSpacing = 8
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 8, bottom: 0, trailing: 8)
+        section.orthogonalScrollingBehavior = .continuous
+
+        return section
+    }
+}
